refactor(registrar): Phase 2 — ?view= deprecation + WelcomeView extraction (#1783)
* refactor(registrar): Phase 2 — deprecate ?view= query param for canonical paths

Migrates the registrar panel from URL query param routing (?view=welcome, ?view=queue) to canonical React Router paths (/registrar/welcome, /registrar/queue), completing audit §5.1 / §8 Strategic Direction 3.

Changes:
- HeaderNew.jsx (2 sites): navigate(`${registrarHomeRoute}?view=welcome|queue`) → navigate('/registrar/welcome|queue')
- Header.jsx (2 sites): same migration as HeaderNew
- RegistrarPanel.jsx breadcrumb: setSearchParams({view:'welcome', ...}) → navigate('/registrar/welcome', {replace:true}) preserving other params
- RegistrarPanel.jsx: added Phase 2 redirect useEffect — when user lands on bare /registrar with ?view=welcome|queue, automatically redirects (replace, no history pollution) to /registrar/welcome|queue preserving all other query params

Backward compat: ?view=welcome|queue still works (the redirect converts it on first load). currentView useMemo still falls back to ?view= if path detection fails. All other params (q, status, date, patientId, dept) unchanged.

Tests: 467/467 passing.

* refactor(registrar): extract WelcomeView from RegistrarPanel (−394 lines)

Extracts the inline welcome dashboard view (lines 1586-2020 of
RegistrarPanel.jsx) into a dedicated src/pages/registrar/views/WelcomeView.jsx
component, mirroring the QueueView.jsx extraction pattern (decomp 6a → 6b).

The welcome view was tightly coupled to ~34 parent scope items: translation
fn (t), language, theme, textColor, appointments, departmentStats,
dataSource, appointmentsLoading, filteredAppointments, services, activeTab,
historyDate, showCalendar, tempDateInput, loadAppointments, setShowWizard,
setWizardEditMode, setWizardInitialData, setShowPaymentManager,
setHistoryDate, setShowCalendar, setTempDateInput, setSearchParams,
navigate, setPaymentDialog, setPrintDialog, setContextMenu,
openRecordPreview, openRecordEditor, updateAppointmentStatus,
handleStartVisit, plus inline-defined helpers generateCSV, downloadCSV and
the memoized DataSourceIndicator component — all now passed as explicit
props.

Result: RegistrarPanel.jsx shrinks 2700 → 2306 lines (−394). The welcome
view's JSX, classNames, aria-labels, event handlers, and structure are
preserved verbatim. Tests: 467/467 passing.

* chore(registrar): cleanup unused imports after WelcomeView extraction

Removes 9 unused imports/vars from RegistrarPanel.jsx that became orphaned after the WelcomeView extraction:
- Card, CardHeader, CardContent, Input (UI components)
- AnimatedTransition (now used only inside WelcomeView)
- ModernQueueManager, ModernStatistics (now used only inside WelcomeView/QueueView)
- getYesterdayDateString (only used by WelcomeView)
- cardBg, borderColor, accentColor (only used by WelcomeView)
- isDark (only used to compute removed cardBg/borderColor)

Pre-existing warnings (left as-is): exhaustive-deps on line 825, singlequote on line 1389.

Tests: 467/467 passing. Lint: 0 errors, 2 pre-existing warnings.

---------
